Conflicts:
	app/controllers/projects/wikis_controller.rb
	app/views/projects/wikis/edit.html.haml
	spec/features/projects/wiki/user_updates_wiki_page_spec.rb

+++ db/migrate/20161212142807_add_lower_path_index_to_routes.rb
@@ -1,6 +1,7 @@
 # See http://doc.gitlab.com/ce/development/migration_style_guide.html
 # for more information on how to write migrations for GitLab.
 
+# rubocop:disable RemoveIndex
 class AddLowerPathIndexToRoutes < ActiveRecord::Migration
   include Gitlab::Database::MigrationHelpers
 
@@ -17,6 +18,6 @@ class AddLowerPathIndexToRoutes < ActiveRecord::Migration
   def down
     return unless Gitlab::Database.postgresql?
 
-    remove_index :routes, name: :index_on_routes_lower_path
+    remove_index :routes, name: :index_on_routes_lower_path if index_exists?(:routes, name: :index_on_routes_lower_path)
   end
 end
